+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [Unreleased]
 ### Added
 - Support the `hasPragma` function.
+- Support the new `number_arg` node type in Ruby 2.7.
 
 ## [0.8.0] - 2019-03-08
 ### Added