+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.9.0] - 2019-03-18
 ### Added
 - Support the `hasPragma` function.
 - Support the new `number_arg` node type in Ruby 2.7.
@@ -208,7 +210,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 - Initial release 🎉
 
-[Unreleased]: https://github.com/prettier/plugin-ruby/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/prettier/plugin-ruby/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/prettier/plugin-ruby/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/prettier/plugin-ruby/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/prettier/plugin-ruby/compare/v0.6.3...v0.7.0
 [0.6.3]: https://github.com/prettier/plugin-ruby/compare/v0.6.2...v0.6.3